clarity around aws secrets manager implementation
customer ran into several days of troubleshooting due to lack of clarity around leading `/` in variable names - I am udpating the `AIRFLOW__SECRETS__BACKEND_KWARGS` to match this, as the prefix setting doesn't discern between using leading `/` or not, but setting the actual secret names will not work if a leading `/` is used.  

Also tried to clarify a few points but unsure if wording or formatting is most ideal. Definitely open to other suggestions, and if some of these are approved we may want to propagate the same changes to the other providers