Create external link checking GitHub action
Lychee seems to have the support and options we need to make a truly useful external broken link checker. I also think it makes sense to run this as a cron job rather than as a PR check - it's rare to ship a broken link in new content and we don't want ancient links to prevent PRs from merging.