Make instructions for private registry cred request more generic
The current set of instructions is not really detailed enough to help a customer provide the necessary information that the support agent needs to complete the request. Ideally, once a request comes in, the support agent will provide the full set of instructions for creating and sending the credentials.

I am not opposed to providing the full instructions on the docs site. The only concern is that the instructions may be too detailed and may bloat the page.